Asynchronous waiting on wait handle
